requirements.txt
certifi==2019.9.11
chardet==3.0.4
dj-database-url==0.5.0
Django==2.2.7
django-heroku==0.3.1
djangorestframework==3.10.3
gunicorn==20.0.0
idna==2.7
psycopg2==2.8.4
pytz==2019.3
requests==2.19.1
sqlparse==0.3.0
urllib3==1.23
whitenoise==4.1.4
